Fixed typos and style, added some remarks
I've added a remark on `using` statement with a `IDisposable` being a return value of a method - while this is not strictly related to code style, I think it fits neatly in the section.

In the section about the switch clause, I've added a word on recursive patterns. Weirdly, though this feature is present since C# 8, the documentation is silent about it. The most comprehensive source describing the feature is this post by Mads Torgersen: https://devblogs.microsoft.com/dotnet/do-more-with-patterns-in-c-8-0/