Fix null conditional example
For the null conditional example, the 'return something?.Other?.Whatever' statement suggests that Other is a property of something and Whatever is a property of Other. In that case the in the bad code block, the container object should be added with the '.' operator. Also, following official guidelines, public properties should start with a Capital letter, so that is fixed to.